requirements.txt
# PyTorch with CUDA 13.0 (install separately)
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu130  OR pip3 install torch torchvision

transformers
Pillow
numpy
tqdm
scikit-learn
matplotlib
streamlit

#pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121 && pip install diffusers==0.27.2 transformers==4.41.2 huggingface_hub==0.23.0 accelerate==1.12.0 safetensors==0.7.0 pillow==11.3.0 streamlit==1.51.0 xformers==0.0.26.post1

#pip install torch==2.5.1+cpu torchvision==0.20.1+cpu torchaudio==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu && pip install diffusers==0.27.2 transformers==4.41.2 huggingface_hub==0.23.0 accelerate==1.12.0 safetensors==0.7.0 pillow==11.3.0 streamlit==1.51.0
